refactor: split blocks.py into modules and update project docs
