+++ CLAUDE.md
@@ -12,7 +12,7 @@ python manage.py runserver
 python manage.py test
 
 # Run a single test module
-python manage.py test blowcomotion.tests.test_member_auth_views
+python manage.py test members.tests.test_member_auth_views
 
 # Check / fix import sorting (runs automatically on commit via pre-commit hook)
 isort --check-only blowcomotion/ --diff
@@ -42,19 +42,19 @@ python manage.py collectstatic --noinput
 
 **Django + Wagtail CMS.** Wagtail handles all CMS pages via a single catch-all URL at the bottom of `urls.py`. Everything above it is a named Django view.
 
-**Single main app:** `blowcomotion/` contains models, views, forms, blocks, templates, and static files. There is no feature-per-app split.
+**Hybrid app layout:** domain apps (`gigs`, `attendance`, `charts`, `instruments`, `members`) hold views, forms, urls, templates, tests, and management commands for their domain. ALL models live in the `blowcomotion` app, in the `blowcomotion/models/` package — new models for any domain still go there, and new apps must not gain models or migrations of their own. `blowcomotion/` itself also keeps StreamField blocks (`blowcomotion/blocks/`), the Wagtail admin wiring (`snippet_viewsets.py`, `chooser_viewsets.py`, `wagtail_hooks.py`), and cross-domain/infra management commands.
 
 **Public form pipeline:** All public-facing form submissions (contact, join band, booking, donate, feedback, member signup) POST to a single `/process-form/` endpoint handled by `process_form()` in `views.py`. The form type is identified by a hidden `form_type` field.
 
-**Member portal** is a separate set of views in `member_views.py` with its own URLs in `member_urls.py`, mounted at `/member/`. It uses Django's built-in auth views as base classes.
+**Member portal** is a separate set of views in `members/views.py` with its own URLs in `members/urls.py`, mounted at `/member/`. It uses Django's built-in auth views as base classes.
 
 **Attendance and birthdays** are protected by HTTP Basic Auth via a custom `@http_basic_auth` decorator that reads passwords from `SiteSettings` (Wagtail admin > Settings).
 
-**StreamField blocks** are defined in `blocks.py` and rendered by templates in `templates/blocks/`. New page content types are added as blocks on `BlankCanvasPage`.
+**StreamField blocks** are defined in the `blowcomotion/blocks/` package (`layout.py`, `content.py`, `forms.py`, `media.py`, re-exported from `blowcomotion/blocks/__init__.py`) and rendered by templates in `templates/blocks/`. New page content types are added as blocks on `BlankCanvasPage`.
 
 **GigoGig integration:** `CachedGig` is a DB table that stores gig data synced from the API (persists until next sync). The `gigs_for_date` view adds a second layer via Django's `cache` (600s TTL) on top of DB queries. Configured via `GIGO_API_URL` / `GIGO_API_KEY` / `GIGO_BAND_ID` in `local.py`.
 
-**Patreon validation:** After an instrument rental request is submitted, `blowcomotion/patreon_client.py` paginates the Patreon API v2 campaign member list to check whether the submitter's email has an active Patreon pledge. The result is stored on `InstrumentRentalRequestSubmission.patreon_validated` (None = not checked, True = active, False = not found / inactive) and included in the manager notification email. Requires in `local.py`:
+**Patreon validation:** After an instrument rental request is submitted, `instruments/patreon.py` paginates the Patreon API v2 campaign member list to check whether the submitter's email has an active Patreon pledge. The result is stored on `InstrumentRentalRequestSubmission.patreon_validated` (None = not checked, True = active, False = not found / inactive) and included in the manager notification email. Requires in `local.py`:
 - `PATREON_ACCESS_TOKEN` — creator access token (must have `campaigns.members` and `campaigns.members[email]` scopes)
 - `PATREON_CAMPAIGN_ID` — numeric campaign ID
 
@@ -63,7 +63,7 @@ If either setting is absent the check is skipped silently (field stays None).
 ## reCAPTCHA — required on every public form
 
 **Every form that accepts public POST submissions must:**
-1. Be validated server-side by `_validate_recaptcha(request)` (in `views.py:525`) before processing
+1. Be validated server-side by `_validate_recaptcha(request)` (in `blowcomotion/views.py`) before processing
 2. Show the disclosure notice "This site is protected by reCAPTCHA." below the submit button
 
 The disclosure notice is injected automatically by `form.js` for any form matching:
